Package-ify FLA and move scripts out of lib
