+++ flexible_load_analysis/analysis/interactive_analysis.py
