+++ flexible_load_analysis/analysis/methods/coincidence_factors.py
