+++ flexible_load_analysis/objects/timeseries.py
